Unwrap the original Exemplars class
Now that the callee needs to handle Unicode Sets,
the handling of Unicode Sets can be consolidated.

The order of the list of graphemes changed when the
unwrap happened, not sure why that happened. Maybe
because Counter objects are not stable in the order
they return items for for items that have the same count.