Remove Unicode Set syntax from test data
The callee will still handle Unicode Sets, the difference is the tests
are simpler without the Unicode Set syntax, and will not have to change
if the Unicode Set generation code changes.

This way the unit test for the Exemplar() class tests just the find
exemplar algorithm, and not the find algorithm plus the Unicode Set
code.

A separate unit test for the Unicode Set code should be added.