GHA: Test if deploy job really needs to checkout sldr repo.
Since the deploy job is rsync-ing data only from the artefact it downloads does
it really need to checkout the sldr repo at all?
Ensure rsync is doing a --dry-run as this is a test: We don't want to delete
the data off the server if not checking out breaks artefact download and
unpacking for any reason.